[Triton-MLIR][BACKEND] Refactor TritonGPUToLLVM into several files (#988)
Refactor the backend into multiple smaller files.
