+++ unittest/Conversion/TritonGPUToLLVM/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_triton_ut(
 	NAME TestPtxAsmFormat
-	SRCS PtxAsmFormatTest.cpp
+	SRCS PTXAsmFormatTest.cpp
 	LIBS TritonGPUToLLVM
 )
